Criando upload de foto do perfil e service do palestrante

+++ Back/src/ProEventos.API/helpers/Util.cs
@@ -22,8 +22,11 @@ namespace ProEventos.API.helpers
         }
         public void DeleteImage(string imageName, string destino)
         {
-            var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, @$"Resources/{destino}", imageName);
-            if (System.IO.File.Exists(imagePath)) System.IO.File.Delete(imagePath);
+            if(!string.IsNullOrEmpty(imageName))
+            {
+                var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, @$"Resources/{destino}", imageName);
+                if (System.IO.File.Exists(imagePath)) System.IO.File.Delete(imagePath);
+            }
         }
     }
 }
